ref: moved all components to separate files
