+++ radio-songs-app/src/components/Header.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+function Header() {
+    return (
+        <div className='header'>
+            <h1>Заявки на Радио</h1>
+            <p>Отправляйте заявку и мы сыграем вашу любимую песню!</p>
+        </div>
+    );
+}
+
+export default Header
